+++ packages/dx-react-chart-demos/src/demo-sources/line-chart/material-ui/step-line.jsx
@@ -24,7 +24,6 @@ const Line = props => (
   />
 );
 
-
 const legendStyles = () => ({
   root: {
     display: 'flex',