+++ packages/dx-react-grid-material-ui/src/plugins/export-panel.jsx
@@ -4,7 +4,6 @@ import { ToggleButton } from '../templates/export-panel/toggle-button';
 import { Menu } from '../templates/export-panel/menu';
 import { MenuItem } from '../templates/export-panel/menu-item';
 
-
 export const ExportPanel = withComponents({
   ToggleButton, Menu, MenuItem,
 })(ExportPanelBase);