+++ packages/dx-react-scheduler-demos/src/demo-sources/scheduler-featured-redux-integration/material-ui/demo.jsx
@@ -261,7 +261,6 @@ const SchedulerContainer = ({
   </Paper>
 );
 
-
 const schedulerInitialState = {
   data: appointments,
   currentDate: '2018-06-27',