+++ packages/dx-site/src/components/layouts/page-layout.jsx
@@ -1,4 +1,3 @@
-
 import * as React from 'react';
 import * as PropTypes from 'prop-types';
 import Layout from './layout';
@@ -6,6 +5,7 @@ import Header from '../header';
 import VersionLink from '../docs/version-link';
 import LandingProductLinks from '../landing/product-links';
 import ContentContainer from './content-container';
+// import SurveyHeaderAddon from '../docs/survey-header-addon';
 
 class PageLayout extends React.PureComponent {
   static propTypes = {
@@ -29,6 +29,7 @@ class PageLayout extends React.PureComponent {
               <LandingProductLinks />
             </>
           )}
+          // addon={<SurveyHeaderAddon />}
         />
         <ContentContainer>
           {children}