+++ CHANGELOG.md
@@ -1,3 +1,41 @@
+# [2.7.6](https://github.com/DevExpress/devextreme-reactive/compare/v2.7.5...v2.7.6) (2021-06-24)
+
+
+### Bug Fixes
+
+* **react-chart:** fix chart display for zoom factor 90% ([#3354](https://github.com/DevExpress/devextreme-reactive/issues/3354)) ([ba7097e](https://github.com/DevExpress/devextreme-reactive/commit/ba7097e3f57b53528b5fa6e9a1857227193c3724)), closes [#2988](https://github.com/DevExpress/devextreme-reactive/issues/2988), [#3288](https://github.com/DevExpress/devextreme-reactive/issues/3288)
+* **react-chart:** fix series showing if scaleName contains `top` or `bottom` ([#3343](https://github.com/DevExpress/devextreme-reactive/issues/3343)) ([99c4efd](https://github.com/DevExpress/devextreme-reactive/commit/99c4efd4b287cc0bb9a84ea49fe7272cd98e75bd))
+* **react-grid:** make possible to define custom messages for TableFilterRow with TypeScript ([#3298](https://github.com/DevExpress/devextreme-reactive/issues/3298)) ([9b08ce5](https://github.com/DevExpress/devextreme-reactive/commit/9b08ce5f1bd1f61ff8bfd7daa3450b0dcbc67050))
+* **react-grid:** show page button instead of ellipsis if all pagination elements are visible ([#3353](https://github.com/DevExpress/devextreme-reactive/issues/3353)) ([c0ee6fe](https://github.com/DevExpress/devextreme-reactive/commit/c0ee6fea7d9fce6763ea7ff92f41ee207d6ca65f))
+
+
+
+# [2.7.5](https://github.com/DevExpress/devextreme-reactive/compare/v2.7.4...v2.7.5) (2021-03-04)
+
+
+### Bug Fixes
+
+* **react-grid:** make `onValueChange` property for TableEditCell optional ([#3253](https://github.com/DevExpress/devextreme-reactive/issues/3253)) ([5c1247f](https://github.com/DevExpress/devextreme-reactive/commit/5c1247f406d4e2b32107cba578db24f14182a229))
+* **react-grid:** take into account flex columns for colspan calculating ([#3254](https://github.com/DevExpress/devextreme-reactive/issues/3254)) ([3f35212](https://github.com/DevExpress/devextreme-reactive/commit/3f35212e362c37d95c794bebee5004653a2bc645))
+* **react-scheduler-material-ui:** show group panel in Mozilla Firefox correctly ([#3264](https://github.com/DevExpress/devextreme-reactive/issues/3264)) ([92576d7](https://github.com/DevExpress/devextreme-reactive/commit/92576d7e4baf18e30c8fe8810978ff404899a0eb))
+
+
+
+# [2.7.4](https://github.com/DevExpress/devextreme-reactive/compare/v2.7.3...v2.7.4) (2020-12-24)
+
+
+### Bug Fixes
+
+* **react-grid:** correct colspan calculating for group row when used with certain plugins ([#3170](https://github.com/DevExpress/devextreme-reactive/issues/3170)) ([c73dd9f](https://github.com/DevExpress/devextreme-reactive/commit/c73dd9f9b72626d99a3db8f12357eeb986464590))
+* **react-grid:** fix regression after [#3174](https://github.com/DevExpress/devextreme-reactive/issues/3174) ([#3214](https://github.com/DevExpress/devextreme-reactive/issues/3214)) ([7ccd25c](https://github.com/DevExpress/devextreme-reactive/commit/7ccd25cfad0c7e44cbbdc086914ac0a4167769d2))
+* **react-grid:** format currency in the Data Editing demo ([#3179](https://github.com/DevExpress/devextreme-reactive/issues/3179)) ([e7ce15f](https://github.com/DevExpress/devextreme-reactive/commit/e7ce15fdef3616f46dbd0700f9cd3b67896f50a2))
+* **react-grid:** make grouping work with a high level of nested groups ([#3174](https://github.com/DevExpress/devextreme-reactive/issues/3174)) ([abb2983](https://github.com/DevExpress/devextreme-reactive/commit/abb29832c95fafb670f02a473947246a651a8919))
+* **react-grid:** make inline cell editing works only with data rows ([#3201](https://github.com/DevExpress/devextreme-reactive/issues/3201)) ([087eb69](https://github.com/DevExpress/devextreme-reactive/commit/087eb69a66f4acef145bd84d0cc554c73f9605df))
+* **react-grid:** make tree data work with a high level of row nesting ([#3171](https://github.com/DevExpress/devextreme-reactive/issues/3171)) ([623883d](https://github.com/DevExpress/devextreme-reactive/commit/623883da813b0969d1139d70cbe7a4069e279f9b))
+* **react-scheduler:** display time in Time Scale correctly when a DST change is present ([#3198](https://github.com/DevExpress/devextreme-reactive/issues/3198)) ([d1247fb](https://github.com/DevExpress/devextreme-reactive/commit/d1247fb196a58d08ba44cd88894cccf5997d6f9e))
+
+
+
 # [2.7.3](https://github.com/DevExpress/devextreme-reactive/compare/v2.7.2...v2.7.3) (2020-11-03)